CSC484project/requirements.txt
Django==5.0.3
mysql-connector-python==8.3.0
django-cors-headers==3.7.0
bootstrap==5.0
mysqlclient==2.2.4

// install node.js and npm axios
